# backend/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FeedTrack API",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://feedtrack-backend.vercel.app",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/login": {
      "post": {
        "summary": "User login",
        "description": "Endpoint for user authentication and login",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User credentials",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "description": "User email address"
                },
                "number": {
                  "type": "string",
                  "description": "User mobile number"
                },
                "password": {
                  "type": "string",
                  "description": "User password"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully logged in",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "User successfully logged in"
                },
                "user": {
                  "$ref": "#/definitions/User"
                },
                "token": {
                  "type": "string",
                  "description": "JWT token for user authentication"
                },
                "secret": {
                  "type": "object",
                  "description": "2FA secret for generating QR code",
                  "properties": {
                    "ascii": {
                      "type": "string",
                      "description": "ASCII representation of the secret"
                    },
                    "hex": {
                      "type": "string",
                      "description": "Hexadecimal representation of the secret"
                    },
                    "base32": {
                      "type": "string",
                      "description": "Base32 representation of the secret"
                    },
                    "otpauth_url": {
                      "type": "string",
                      "description": "URL for OTP authentication setup"
                    }
                  }
                },
                "verified": {
                  "type": "boolean",
                  "description": "Flag indicating if user is verified"
                }
              }
            }
          },
          "400": {
            "description": "Invalid email or mobile number, or incorrect password",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "enum": ["Email or mobile number incorrect!", "Email or mobile number is required!", "Password incorrect!"]
                        }
                      }
                    },
                    {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "Email or mobile number incorrect!"
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/googleAddUser": {
      "post": {
        "summary": "Add user via Google",
        "description": "Endpoint for adding a new user to the database using Google authentication",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "description": "User ID"
                },
                "name": {
                  "type": "string",
                  "description": "User first name"
                },
                "lastname": {
                  "type": "string",
                  "description": "User last name"
                },
                "email": {
                  "type": "string",
                  "description": "User email address"
                },
                "image": {
                  "type": "string",
                  "description": "User profile image URL"
                },
                "password": {
                  "type": "string",
                  "description": "User password"
                },
                "mobilenumber": {
                  "type": "string",
                  "description": "User mobile number"
                },
                "role": {
                  "type": "string",
                  "description": "User role (optional, default: superAdmin)"
                },
                "verified": {
                  "type": "boolean",
                  "description": "Flag indicating if user is verified"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User successfully added",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "User added successfully"
                },
                "user": {
                  "$ref": "#/definitions/User"
                },
                "token": {
                  "type": "string",
                  "description": "JWT token for user authentication"
                },
                "secret": {
                  "type": "object",
                  "description": "2FA secret for generating QR code",
                  "properties": {
                    "ascii": {
                      "type": "string",
                      "description": "ASCII representation of the secret"
                    },
                    "hex": {
                      "type": "string",
                      "description": "Hexadecimal representation of the secret"
                    },
                    "base32": {
                      "type": "string",
                      "description": "Base32 representation of the secret"
                    },
                    "otpauth_url": {
                      "type": "string",
                      "description": "URL for OTP authentication setup"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid email address or user already exists",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "enum": ["Invalid email address", "User already exists"]
                        }
                      }
                    },
                    {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "Invalid email address"
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/logout": {
      "post": {
        "summary": "User logout",
        "description": "Endpoint for user logout and token invalidation",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User token",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string",
                  "description": "JWT token for user authentication"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully logged out",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Logged out successfully."
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "You are not authenticated!"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Token is not valid!"
                }
              }
            }
          }
        }
      }
    },
    "/api/2faSetup": {
      "post": {
        "summary": "Setup 2FA",
        "description": "Endpoint for generating data URL for 2FA setup",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "2FA secret",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "secret": {
                  "type": "object",
                  "description": "2FA secret for generating QR code",
                  "properties": {
                    "ascii": {
                      "type": "string",
                      "description": "ASCII representation of the secret"
                    },
                    "hex": {
                      "type": "string",
                      "description": "Hexadecimal representation of the secret"
                    },
                    "base32": {
                      "type": "string",
                      "description": "Base32 representation of the secret"
                    },
                    "otpauth_url": {
                      "type": "string",
                      "description": "URL for OTP authentication setup"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Data URL for 2FA setup generated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "dataUrl": {
                  "type": "string",
                  "description": "Data URL for 2FA setup QR code"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error generating QR code"
                }
              }
            }
          }
        }
      }
    },
    "/api/verify": {
      "post": {
        "summary": "Verify 2FA token",
        "description": "Endpoint for verifying 2FA token",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "2FA token and secret",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userToken": {
                  "type": "string",
                  "description": "2FA token entered by the user"
                },
                "secret": {
                  "type": "object",
                  "description": "2FA secret",
                  "properties": {
                    "base32": {
                      "type": "string",
                      "description": "Base32 representation of the secret"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "2FA token verified successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "description": "Flag indicating if 2FA token is verified"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}